refactor(assert drive): Move private method out of anonymous class PE-679
