+++ src/arfsdao.ts
@@ -3,7 +3,15 @@ import type { JWKWallet, Wallet } from './wallet';
 import Arweave from 'arweave';
 import { v4 as uuidv4 } from 'uuid';
 import Transaction from 'arweave/node/lib/transaction';
-import { deriveDriveKey, GQLEdgeInterface, GQLNodeInterface, GQLTagInterface, JWKInterface } from 'ardrive-core-js';
+import {
+	deriveDriveKey,
+	GQLEdgeInterface,
+	GQLNodeInterface,
+	GQLTagInterface,
+	JWKInterface,
+	DrivePrivacy,
+	driveDecrypt
+} from 'ardrive-core-js';
 import {
 	ArFSPublicFileDataPrototype,
 	ArFSObjectMetadataPrototype,
@@ -906,6 +914,51 @@ export class ArFSDAO extends ArFSDAOAnonymous {
 			folderId,
 			await this.getAllFoldersOfPublicDrive({ driveId, owner, latestRevisionsOnly: true })
 		);
+	}
+
+	public async getOwnerAndAssertDrive(driveId: DriveID, driveKey?: DriveKey): Promise<ArweaveAddress> {
+		const gqlQuery = buildQuery({ tags: [{ name: 'Drive-Id', value: `${driveId}` }], sort: ASCENDING_ORDER });
+		const response = await this.arweave.api.post(graphQLURL, gqlQuery);
+		const edges: GQLEdgeInterface[] = response.data.data.transactions.edges;
+
+		if (!edges.length) {
+			throw new Error(`Could not find a transaction with "Drive-Id": ${driveId}`);
+		}
+
+		const edgeOfFirstDrive = edges[0];
+
+		const drivePrivacy: DrivePrivacy = driveKey ? 'private' : 'public';
+		const drivePrivacyFromTag = edgeOfFirstDrive.node.tags.find((t) => t.name === 'Drive-Privacy');
+
+		if (!drivePrivacyFromTag) {
+			throw new Error('Target drive has no "Drive-Privacy" tag!');
+		}
+
+		if (drivePrivacyFromTag.value !== drivePrivacy) {
+			throw new Error(`Target drive is not a ${drivePrivacy} drive!`);
+		}
+
+		if (driveKey) {
+			const cipherIVFromTag = edgeOfFirstDrive.node.tags.find((t) => t.name === 'Cipher-IV');
+			if (!cipherIVFromTag) {
+				throw new Error('Target private drive has no "Cipher-IV" tag!');
+			}
+
+			const driveDataBuffer = Buffer.from(
+				await this.arweave.transactions.getData(edgeOfFirstDrive.node.id, { decode: true })
+			);
+
+			try {
+				// Attempt to decrypt drive to assert drive key is correct
+				await driveDecrypt(cipherIVFromTag.value, driveKey, driveDataBuffer);
+			} catch {
+				throw new Error('Provided drive key or password could not decrypt target private drive!');
+			}
+		}
+
+		const driveOwnerAddress = edgeOfFirstDrive.node.owner.address;
+
+		return new ArweaveAddress(driveOwnerAddress);
 	}
 
 	/**
